+++ tests/class.global.test.mjs
@@ -10,4 +10,5 @@ test('class.global', () => {
     expect(property_exists).not.toBeUndefined();
     expect(get_object_vars).not.toBeUndefined();
     expect(get_class_vars).not.toBeUndefined();
+    expect(get_class_methods).not.toBeUndefined();
 });